+++ R/ABCDscores-package.R
@@ -1,4 +1,6 @@
 #' @keywords internal
+#' @references Zhang, L., Celhay, O., Das, B., Berman, S., Ziemer, L. R., Smith, C. J., Dale, A. M., & Linkersdörfer, J. (2025).
+#' ABCDscores: An R package for computing summary scores in the ABCD Study. \emph{bioRxiv}. \doi{10.1101/2025.09.04.674066}
 "_PACKAGE"
 
 ## usethis namespace: start